Adds files
This doesn't mean the project is active again.
In fact, the only reason I did this is because I was bored and wanted to see how quick I could develop a FNAF clone on Godot Engine. Turns out you can build the basics in less than a day.
I will probably not finish this (just as the other ones) but it was a fun experience anyway.